sync-ed serial with socket

+++ src/Odin/socketodin.cpp
@@ -26,7 +26,7 @@ void SocketOdin::readCommand(){
         if(incomingCommand.size() >= 16){
 //            qDebug() << incomingCommand.toHex();
             for(int i = 0; i < incomingCommand.size(); i++){
-                if(incomingCommand.at(i) != 0xAA){
+                if((quint8) incomingCommand.at(i) != (quint8) 0xAA){
                     incomingCommand.remove(0, 1);
                 }
                 else{
